OOP rework

+++ lib/sf.c
@@ -2,7 +2,7 @@
 #include "sf.h"
 
 const unsigned char PROTOCOL_VER = 0;
-const short DEFAULT_PORT = 6000;
+const int DEFAULT_PORT = 6000;
 
 void alloc_buffer(uv_handle_t* handle, size_t suggested_size, uv_buf_t* buf)
 {
